requirements.txt
appdirs==1.3.0
attrs==18.2.0
black==18.9b0
boto3==1.15.8
botocore==1.18.8
certifi==2018.11.29
chardet==3.0.4
click==7.1.2
cloudrail==0.1.213
colorama==0.4.3
dataclasses-json==0.5.2
fake-useragent==0.1.11
Flask==0.12.2
future==0.17.1
halo==0.0.30
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
junit-xml==1.9
log-symbols==0.0.14
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
natsort==3.3.0
oauthlib==3.0.1
python-dateutil==2.8.1
python-twitter==3.5
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
six==1.15.0
soupsieve==1.8
spinners==0.0.24
stringcase==1.2.0
tabulate==0.8.7
termcolor==1.1.0
toml==0.10.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.24.2
Werkzeug==1.0.1
